% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-core.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached downloads}
\usage{
clear_cache()
}
\value{
None.
}
\description{
All data fetched from the OpenF1 API is cached, and subsequent calls with 
the same parameters just returns the cached value.  Use this function 
to clear this cache so that the next download fetches data from the API
}
\examples{
clear_cache()
}
