% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\name{f1_laps}
\alias{f1_laps}
\title{Provides detailed information about individual laps.}
\usage{
f1_laps(
  ...,
  session_key = "latest",
  driver_number,
  duration_sector_1,
  duration_sector_2,
  duration_sector_3,
  i1_speed,
  i2_speed,
  is_pit_out_lap,
  lap_duration,
  lap_number,
  segments_sector_1,
  segments_sector_2,
  segments_sector_3,
  st_speed,
  date_start,
  meeting_key
)
}
\arguments{
\item{...}{Verbatim options added to API call. Use this to add filtering 
options e.g. \code{f1_sessions("date_start >= 2025-04-01")}.  For filtering
\itemize{
  \item{most named arguments can be filtered}
  \item{use \code{<=}, \code{>=}, \code{<}, \code{>} for specifying ranges}
  \item{for date filtering, use date format "YYYY-MM-DD", and
        "YYYY-MM-DDTHH:MM:SS"}
}}

\item{session_key}{The unique identifier for the session. For 
\code{f1_sessions()} the default is \code{NULL} to return all sessions.
For all other functions \code{session_key} defaults to 'latest' so as to
only fetch the data for the latest/current session. Valid values are:
'latest', \code{NULL} or an integer session key}

\item{driver_number}{The unique number assigned to an F1 driver}

\item{duration_sector_1, duration_sector_2, duration_sector_3}{The time taken, 
in seconds to complete the given sector of the lap}

\item{i1_speed, i2_speed}{The speed of the car, in km/h, at the first and
second intermediate point on the track.}

\item{is_pit_out_lap}{A boolean value indicating whether the lap is an 
"out lap" from the pit (true if it is, false otherwise).}

\item{lap_duration}{The total time taken, in seconds, to complete the entire lap.}

\item{lap_number}{The sequential number of the lap within the session
(starts at 1).}

\item{segments_sector_1, segments_sector_2, segments_sector_3}{A list of 
values representing the "mini-sectors" within the first sector}

\item{st_speed}{The speed of the car, in km/h, at the speed trap, which 
is a specific point on the track where the highest speeds are 
usually recorded.}

\item{date_start}{The UTC starting date and time, in ISO 8601 format.}

\item{meeting_key}{The unique identifier for the meeting. Use 'latest' to
identify the latest or current meeting.}
}
\value{
data.frame
}
\description{
Provides detailed information about individual laps.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
f1_laps(driver_number = 44, lap_number = 1)
\dontshow{\}) # examplesIf}
}
