% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\name{f1_locations}
\alias{f1_locations}
\title{The approximate location of the cars on the circuit, at a sample rate of 
about 3.7 Hz.}
\usage{
f1_locations(
  ...,
  session_key = "latest",
  driver_number,
  x,
  y,
  z,
  date,
  meeting_key
)
}
\arguments{
\item{...}{Verbatim options added to API call. Use this to add filtering 
options e.g. \code{f1_sessions("date_start >= 2025-04-01")}.  For filtering
\itemize{
  \item{most named arguments can be filtered}
  \item{use \code{<=}, \code{>=}, \code{<}, \code{>} for specifying ranges}
  \item{for date filtering, use date format "YYYY-MM-DD", and
        "YYYY-MM-DDTHH:MM:SS"}
}}

\item{session_key}{The unique identifier for the session. For 
\code{f1_sessions()} the default is \code{NULL} to return all sessions.
For all other functions \code{session_key} defaults to 'latest' so as to
only fetch the data for the latest/current session. Valid values are:
'latest', \code{NULL} or an integer session key}

\item{driver_number}{The unique number assigned to an F1 driver}

\item{x, y, z}{3D Cartesian coordinate system representing the current 
approximate location of the car on the track.}

\item{date}{The UTC date and time, in ISO 8601 format.}

\item{meeting_key}{The unique identifier for the meeting. Use 'latest' to
identify the latest or current meeting.}
}
\value{
data.frame
}
\description{
Useful for gauging their progress along the track, but lacks details about 
lateral placement — i.e. whether the car is on the left or right side of the 
track. The origin point (0, 0, 0) appears to be arbitrary and not tied to any 
specific location on the track.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
f1_locations(driver_number = 44)
\dontshow{\}) # examplesIf}
}
