% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\name{f1_meetings}
\alias{f1_meetings}
\title{Provides information about meetings i.e. a Grand Prix or testing weekend.
Usually includes multiple sessions (practice, qualifying, race, ...).}
\usage{
f1_meetings(
  ...,
  circuit_key,
  circuit_short_name,
  country_code,
  country_key,
  country_name,
  date_start,
  gmt_offset,
  location,
  meeting_key,
  meeting_name,
  meeting_official_name,
  year
)
}
\arguments{
\item{...}{Verbatim options added to API call. Use this to add filtering 
options e.g. \code{f1_sessions("date_start >= 2025-04-01")}.  For filtering
\itemize{
  \item{most named arguments can be filtered}
  \item{use \code{<=}, \code{>=}, \code{<}, \code{>} for specifying ranges}
  \item{for date filtering, use date format "YYYY-MM-DD", and
        "YYYY-MM-DDTHH:MM:SS"}
}}

\item{circuit_key}{The unique identifier for the circuit where the event 
takes place.}

\item{circuit_short_name}{The short or common name of the circuit where the 
event takes place.}

\item{country_code}{A code that uniquely identifies the country. e.g. "AUS", "USA"}

\item{country_key}{The unique identifier for the country where the event 
takes place.}

\item{country_name}{The full name of the country where the event takes place.}

\item{date_start}{The UTC starting date and time, in ISO 8601 format.}

\item{gmt_offset}{The difference in hours and minutes between local time at 
the location of the event and Greenwich Mean Time (GMT).}

\item{location}{The city or geographical location where the event takes place.}

\item{meeting_key}{The unique identifier for the meeting. Use 'latest' to
identify the latest or current meeting.}

\item{meeting_name}{The name of the meeting.}

\item{meeting_official_name}{The official name of the meeting.}

\item{year}{The year the event takes place.}
}
\value{
data.frame
}
\description{
Provides information about meetings i.e. a Grand Prix or testing weekend.
Usually includes multiple sessions (practice, qualifying, race, ...).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
f1_meetings()
\dontshow{\}) # examplesIf}
}
