% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\name{f1_race_control}
\alias{f1_race_control}
\title{Provides information about race control (racing incidents, flags, 
safety car, ...).}
\usage{
f1_race_control(
  ...,
  session_key = "latest",
  category,
  driver_number,
  flag,
  lap_number,
  message,
  scope,
  sector,
  date,
  meeting_key
)
}
\arguments{
\item{...}{Verbatim options added to API call. Use this to add filtering 
options e.g. \code{f1_sessions("date_start >= 2025-04-01")}.  For filtering
\itemize{
  \item{most named arguments can be filtered}
  \item{use \code{<=}, \code{>=}, \code{<}, \code{>} for specifying ranges}
  \item{for date filtering, use date format "YYYY-MM-DD", and
        "YYYY-MM-DDTHH:MM:SS"}
}}

\item{session_key}{The unique identifier for the session. For 
\code{f1_sessions()} the default is \code{NULL} to return all sessions.
For all other functions \code{session_key} defaults to 'latest' so as to
only fetch the data for the latest/current session. Valid values are:
'latest', \code{NULL} or an integer session key}

\item{category}{The category of the event 
('CarEvent', 'Drs', 'Flag', 'SafetyCar', ...).}

\item{driver_number}{The unique number assigned to an F1 driver}

\item{flag}{Type of flag displayed 
('GREEN', 'YELLOW', 'DOUBLE YELLOW', 'CHEQUERED', 'RED", 'CLEAR',}

\item{lap_number}{The sequential number of the lap within the session
(starts at 1).}

\item{message}{Description of the event or action.}

\item{scope}{The scope of the event ('Track', 'Driver', 'Sector', ...).}

\item{sector}{Segment ("mini-sector") of the track where the event occurred? 
(starts at 1).}

\item{date}{The UTC date and time, in ISO 8601 format.}

\item{meeting_key}{The unique identifier for the meeting. Use 'latest' to
identify the latest or current meeting.}
}
\value{
data.frame
}
\description{
Provides information about race control (racing incidents, flags, 
safety car, ...).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
f1_race_control()
\dontshow{\}) # examplesIf}
}
