% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\name{f1_sessions}
\alias{f1_sessions}
\title{Provides information about sessions - i.e. a distinct period of track 
activity during a Grand Prix  (practice, qualifying, sprint, race, ...).}
\usage{
f1_sessions(
  ...,
  session_key = NULL,
  circuit_key,
  circuit_short_name,
  country_code,
  country_key,
  country_name,
  date_start,
  date_end,
  gmt_offset,
  location,
  meeting_key,
  session_name,
  session_type,
  year
)
}
\arguments{
\item{...}{Verbatim options added to API call. Use this to add filtering 
options e.g. \code{f1_sessions("date_start >= 2025-04-01")}.  For filtering
\itemize{
  \item{most named arguments can be filtered}
  \item{use \code{<=}, \code{>=}, \code{<}, \code{>} for specifying ranges}
  \item{for date filtering, use date format "YYYY-MM-DD", and
        "YYYY-MM-DDTHH:MM:SS"}
}}

\item{session_key}{The unique identifier for the session. For 
\code{f1_sessions()} the default is \code{NULL} to return all sessions.
For all other functions \code{session_key} defaults to 'latest' so as to
only fetch the data for the latest/current session. Valid values are:
'latest', \code{NULL} or an integer session key}

\item{circuit_key}{The unique identifier for the circuit where the event 
takes place.}

\item{circuit_short_name}{The short or common name of the circuit where the 
event takes place.}

\item{country_code}{A code that uniquely identifies the country. e.g. "AUS", "USA"}

\item{country_key}{The unique identifier for the country where the event 
takes place.}

\item{country_name}{The full name of the country where the event takes place.}

\item{date_start}{The UTC starting date and time, in ISO 8601 format.}

\item{date_end}{The UTC ending date and time, in ISO 8601 format.}

\item{gmt_offset}{The difference in hours and minutes between local time at 
the location of the event and Greenwich Mean Time (GMT).}

\item{location}{The city or geographical location where the event takes place.}

\item{meeting_key}{The unique identifier for the meeting. Use 'latest' to
identify the latest or current meeting.}

\item{session_name}{The name of the session 'Practice 1', 'Qualifying', 
'Race' etc}

\item{session_type}{The type of the session 'Practice', 'Qualifying', 
'Race' etc.}

\item{year}{The year the event takes place.}
}
\value{
data.frame
}
\description{
Provides information about sessions - i.e. a distinct period of track 
activity during a Grand Prix  (practice, qualifying, sprint, race, ...).
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Find all sessions in the OpenF1 database
f1_sessions()[1:5] |>
  head()

# Show all sessions in the last 50 days
# This will user filtering by passing in an unnamed character string 
limit <- Sys.Date() - 50
limit <- sprintf("date_start > \%s", limit)
limit
f1_sessions(limit, session_type = 'Race')[,1:5]
\dontshow{\}) # examplesIf}
}
