% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\name{f1_stints}
\alias{f1_stints}
\title{Provides information about individual stints i.e. a period of continuous 
driving by a driver during a session.}
\usage{
f1_stints(
  ...,
  session_key = "latest",
  driver_number,
  compound,
  lap_end,
  lap_start,
  stint_number,
  tyre_age_at_start,
  meeting_key
)
}
\arguments{
\item{...}{Verbatim options added to API call. Use this to add filtering 
options e.g. \code{f1_sessions("date_start >= 2025-04-01")}.  For filtering
\itemize{
  \item{most named arguments can be filtered}
  \item{use \code{<=}, \code{>=}, \code{<}, \code{>} for specifying ranges}
  \item{for date filtering, use date format "YYYY-MM-DD", and
        "YYYY-MM-DDTHH:MM:SS"}
}}

\item{session_key}{The unique identifier for the session. For 
\code{f1_sessions()} the default is \code{NULL} to return all sessions.
For all other functions \code{session_key} defaults to 'latest' so as to
only fetch the data for the latest/current session. Valid values are:
'latest', \code{NULL} or an integer session key}

\item{driver_number}{The unique number assigned to an F1 driver}

\item{compound}{The specific compound of tyre used during the stint 
(SOFT, MEDIUM, HARD, ...).}

\item{lap_start, lap_end}{Number of the first/last completed lap in this stint.}

\item{stint_number}{The sequential number of the stint within the session 
(starts at 1).}

\item{tyre_age_at_start}{The age of the tyres at the start of the stint, 
in laps completed.}

\item{meeting_key}{The unique identifier for the meeting. Use 'latest' to
identify the latest or current meeting.}
}
\value{
data.frame
}
\description{
Provides information about individual stints i.e. a period of continuous 
driving by a driver during a session.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
f1_stints(driver_number = 44)
\dontshow{\}) # examplesIf}
}
