% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openf1r-package.R
\docType{package}
\name{openf1r-package}
\alias{openf1r}
\alias{openf1r-package}
\title{openf1r: Retrieve Formula 1 Race Data from the OpenF1 API}
\description{
An interface to the OpenF1 API to retrieve Formula 1 race data such as race timings, lap speeds, driver information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/openf1r}
  \item Report bugs at \url{https://github.com/coolbutuseless/openf1r/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

}
\keyword{internal}
