% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1-api.R
\docType{data}
\name{segments_sector}
\alias{segments_sector}
\title{Information about the codes for mini-segments}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\usage{
segments_sector
}
\description{
Information about the codes for mini-segments
}
\keyword{datasets}
